Make warmup a function of time rather than a number of iterations.
This changes warmup from a number of iterations to an amount of time waiting
for the engine to warm up. This should help keep warmup times contained with
slow engines and/or commands while still providing enough warmup time to
engines in most cases.

This also switches to a progress bar, as printing each iteration number on a
new line could exhaust my shell window's history, making it challenging to
check the overall progress of the benchmark.