Display 50th, 90th and 99th percentiles in results.
I have been using this benchmark to test Lucene changes recently, and it would
have been useful for me to not only understand how the average latency was
impacted, but also percentiles. This helps reason about some trade-offs that
would make the faster queries slower but the slower queries faster.

I thought it'd be useful in general, hence this contribution. As you can
probably tell, I have _zero_ frontend background, so I'm sure that there are
better ways to do this.