Add filtering commands.
These new commands allow running queries against a filter that matches 1% or
10% of documents.

Filters are interesting because some optimizations that are easy/obvious for
exhaustive evaluation become more complicated when a filter is applied. Yet
filters are common, think of an e-commerce search filtered by category for
instance.

Only the Lucene 10.0 engine supports filtering for now because I'm not too
familiar with Rust, but I assume that it should be easy to add support for it
in a similar fashion.